refactor: BaseTime 위치 수정(DDD 고려)
